break microagent init to multiple files
